# src/main/resources/api.yml
openapi: 3.0.1
info:
  title: Demo
  description: 'Demo'
  version: 0.5.0
servers:
  - url: https://op.op.com
...
paths:
...
  /reporte/{id}:
    get:
      tags:
        - Reporte
      summary: Retorna un Reporte especifico
      operationId: getReporteById
      parameters:
        - name: id
          in: path
          description: ID del reporte a retornar
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Reporte encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReporteResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
    put:
      tags:
        - Reporte
      summary: Actualizar un Reporte
      operationId: updateReporte
      parameters:
        - name: id
          in: path
          description: ID del reporte a editar
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Actualizar un reporte existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReporteRequest'
        required: true
      responses:
        200:
          description: Reporte actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
...
  /asistencia/{id}:
    get:
      tags:
        - Asistencia
      summary: Retorna una Asistencia especifica
      operationId: getAsistenciaById
      parameters:
        - name: id
          in: path
          description: ID de la asistencia a retornar
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Asistencia encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsistenciaResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
    put:
      tags:
        - Asistencia
      summary: Actualizar una Asistencia
      operationId: updateAsistencia
      parameters:
        - name: id
          in: path
          description: ID de la asistencia a editar
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Actualizar una asistencia existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsistenciaRequest'
        required: true
      responses:
        200:
          description: Asistencia actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
    patch:
      tags:
        - Asistencia
      summary: Actualizar el estado de una Asistencia
      operationId: updateEstadoAsistencia
      parameters:
        - name: id
          in: path
          description: ID de la asistencia a editar
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Actualizar el estado de una asistencia existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstadoAsistenciaRequest'
        required: true
      responses:
        200:
          description: Asistencia actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
...
  /tecnico/{id}:
    get:
      tags:
        - Technician
      summary: Retorna un Technician especifico
      operationId: getTechnicianById
      parameters:
        - name: id
          in: path
          description: ID del tecnico a retornar
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Technician encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicianResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
    put:
      tags:
        - Technician
      summary: Actualizar un Technician
      operationId: updateTechnician
      parameters:
        - name: id
          in: path
          description: ID del tecnico a editar
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Actualizar un tecnico existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TechnicianRequest'
        required: true
      responses:
        200:
          description: Asistencia actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
...
components:
  schemas:
...
    ReporteRequest:
      type: object
      required:
        - descripcion
        - longitud
        - latitud
        - telefono
      properties:
        telefono:
          type: string
        descripcion:
          type: string
        tipoReporte:
          type: string
          enum:
            - Hoyo en Pavimento
            - Otro
        longitud:
          type: string
        latitud:
          type: string
        notas:
          type: string
        fotos:
          type: array
          items:
            type: string
    Reporte:
      type: object
      properties:
        id:
          type: integer
          format: int64
        descripcion:
          type: string
        tipoReporte:
          type: string
        longitud:
          type: string
        latitud:
          type: string
        notas:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        correo:
          type: string
        telefono:
          type: string
        fechaCreacion:
          type: string
          format: date-time
        fechaUltActualizacion:
          type: string
          format: date-time
        fotos:
          type: array
          items:
            type: string
        estado:
          $ref: '#/components/schemas/EstadoReporte'
    ReporteResponse:
      type: object
      properties:
        value:
          $ref: '#/components/schemas/Reporte'
        type:
          type: string
        message:
          type: string
...
    AsistenciaRequest:
      type: object
      required:
        - nombre
        - apellido
        - telefono
        - tipoSituacion
        - placa
        - latitud
        - longitud
      properties:
        fechaAsistencia:
          type: string
          format: date-time
        nombre:
          type: string
          maxLength: 100
        apellido:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
        tipoSituacion:
          type: string
          maxLength: 100
        situacion:
          type: string
          maxLength: 2000
        placa:
          type: string
          maxLength: 20
        tipoVehiculo:
          type: string
          maxLength: 50
        marca:
          type: string
          maxLength: 50
        modelo:
          type: string
          maxLength: 50
        color:
          type: string
          maxLength: 20
        urlImagenFirma:
          type: string
          maxLength: 250
        latitud:
          type: string
          maxLength: 50
        longitud:
          type: string
          maxLength: 50
        idTechnician:
          type: integer
          format: int64
        notasTechnician:
          type: string
          maxLength: 2000
        estado:
          type: string
          description: Estado de la asistencia
          enum:
            - Enviada
            - Asignada
            - En Proceso
            - Completada
            - Cerrada
    EstadoAsistenciaRequest:
      type: object
      required:
        - estado
        - notasTechnician
      properties:
        notasTechnician:
          type: string
          maxLength: 2000
        urlImagenFirma:
          type: string
          maxLength: 250
        estado:
          type: string
          description: Estado de la asistencia
          enum:
            - Enviada
            - Asignada
            - En Proceso
            - Completada
            - Cerrada
    Asistencia:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fechaLlamada:
          type: string
          format: date-time
        fechaAsistencia:
          type: string
          format: date-time
        nombre:
          type: string
          maxLength: 100
        apellido:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
        tipoSituacion:
          type: string
          maxLength: 100
        situacion:
          type: string
          maxLength: 2000
        placa:
          type: string
          maxLength: 20
        tipoVehiculo:
          type: string
          maxLength: 50
        marca:
          type: string
          maxLength: 50
        modelo:
          type: string
          maxLength: 50
        color:
          type: string
          maxLength: 20
        urlImagenFirma:
          type: string
          maxLength: 250
        latitud:
          type: string
          maxLength: 50
        longitud:
          type: string
          maxLength: 50
        idTechnician:
          type: integer
          format: int64
        notasTechnician:
          type: string
          maxLength: 2000
        tecnico:
          $ref: '#/components/schemas/Technician'
        estado:
          type: string
          description: Estado de la asistencia
          enum:
            - Enviada
            - Asignada
            - En Proceso
            - Completada
            - Cerrada
    AsistenciaResponse:
      type: object
      properties:
        value:
          $ref: '#/components/schemas/Asistencia'
        type:
          type: string
        message:
          type: string
...
    TechnicianRequest:
      type: object
      required:
        - nombre
        - apellido
        - telefono
        - latitud
        - longitud
      properties:
        nombre:
          type: string
          maxLength: 100
        apellido:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
        latitud:
          type: string
          maxLength: 50
        longitud:
          type: string
          maxLength: 50
    Technician:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
          maxLength: 100
        apellido:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
        latitud:
          type: string
          maxLength: 50
        longitud:
          type: string
          maxLength: 50
    TechnicianResponse:
      type: object
      properties:
        value:
          $ref: '#/components/schemas/Technician'
        type:
          type: string
        message:
          type: string
...
    EstadoReporte:
      type: object
      properties:
        descripcion:
          type: string
          maxLength: 50
        descripcionSecundaria:
          type: string
          maxLength: 50
        porcentajeCompletado:
          type: number
          format: float
        color:
          type: string
        detalle:
          type: array
          items:
            $ref: '#/components/schemas/DetalleEstadoReporte'
    DetalleEstadoReporte:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 50
        fechaInicio:
          type: string
          format: date-time
        fechaFinal:
          type: string
          format: date-time
...
    EmptyResponse:
      type: object
      properties:
        type:
          type: string
        message:
          type: string